pyproject.toml
[project]
name = "yang-ban"
version = "0.1.0"
description = "RisuAI Character Proxy with Observability"
readme = "CLAUDE.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "uvloop>=0.21.0",
    "structlog>=24.4.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "langfuse>=2.51.0",
    "prometheus-client>=0.21.0",
    "starlette-prometheus>=0.10.0",
    "mem0ai>=0.1.40",
    "qdrant-client>=1.12.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "python-multipart>=0.0.12",
    "aiosqlite>=0.20.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
